refactor: 인수테스트 픽스쳐 수정
